[INTERNAL] JSDoc: fail on well known, broken type references
This change switches the logging from a hand made console wrapper to
JSDoc's own logger module. The module implicitly manages the exit code
of JSDoc as as side effect.

For now, the validation only fails for a small set of well-known wrong
type names that are commonly used by developers. The correct type name
is mentioned in a hint, together with the error message.

+++ src/sap.m/src/sap/m/upload/Uploader.js
@@ -70,12 +70,12 @@ sap.ui.define([
 						 * The number of bytes transferred since the beginning of the operation.
 						 * This doesn't include headers and other overhead, but only the content itself
 						 */
-						loaded: {type: "long"},
+						loaded: {type: "int"},
 						/**
 						 * The total number of bytes of content that will be transferred during the operation.
 						 * If the total size is unknown, this value is zero.
 						 */
-						total: {type: "long"}
+						total: {type: "int"}
 					}
 				},
 				/**
